+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   - ValueTypeMap.String() -> "Map"
   - ValueTypeSlice.String() -> "Slice"
   - ValueTypeBytes.String() -> "Bytes"
+- Rename output of `[MetricType|NumberDataPointValueType|ExemplarValueType].String()` for zero values from `"None"` to 
+  `"Empty"` (#6270)
 
 ### 🚩 Deprecations 🚩
 
@@ -36,10 +38,16 @@
 - Deprecate `pcommon.NewValueString` in favor of `pcommon.NewValueStr` (#6209)
 - Deprecate `pmetric.MetricAggregationTemporality` enum type in favor of `pmetric.AggregationTemporality` (#6249)
 - Deprecate `confmap.Conf.UnmarshalExact` in favor of `confmap.Conf.Unmarshal(.., WithErrorUnused)` (#6231)
+- Deprecate `pmetric.MetricDataPointFlags` favor of `pmetric.DataPointFlags` (#6259)
 - Deprecate `ptrace.SpanStatus` favor of `ptrace.Status` (#6258)
 - Deprecate `pmetric.[New]?Buckets` in favor of `pmetric.[New]?ExponentialHistogramDataPointBuckets` (#6261)
 - Deprecate `plog.SeverityNumberUndefined` in favor of `plog.SeverityNumberUnspecified` (#6269)
 - Deprecate `pmetric.[New]?ValueAtQuantile[Slice]?` in favor of `pmetric.[New]?SummaryDataPointValueAtQuantile[Slice]?` (#6262)
+- Deprecate enum zero constants ending with `None` (#6270)
+  - Deprecate `pmetric.MetricTypeNone` in favor of `pmetric.MetricTypeEmpty`
+  - Deprecate `pmetric.NumberDataPointValueTypeNone` in favor of `pmetric.NumberDataPointValueTypeEmpty`
+  - Deprecate `pmetric.ExemplarValueTypeNone` in favor of `pmetric.ExemplarValueTypeEmpty`
+- Deprecate `plog.SeverityNumberUndefined` in favor of `plog.SeverityNumberUnspecified` (#6269)
 
 ### 💡 Enhancements 💡